Rename data structures in sync
