+++ src/sync/mod.rs
@@ -1,4 +1,4 @@
 //! Synchronization primitives.
 
-pub mod michael_list;
-pub mod ms_queue;
+pub mod list;
+pub mod queue;
